fix minus 13ss

+++ printf/includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: kshanti <kshanti@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/21 15:08:11 by kshanti           #+#    #+#             */
-/*   Updated: 2020/11/30 15:52:28 by kshanti          ###   ########.fr       */
+/*   Updated: 2020/11/30 17:38:01 by kshanti          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,11 +31,11 @@ int				column_width(t_arg *tmp, long long num);
 int				column_precision(t_arg *tmp, long long num);
 int				column_width_str(t_arg *tmp, char *str);
 int				output_precision_str(t_arg *tmp, char *str);
-int             precision_poiner(t_arg *tmp, long long n);
+int             precision_poiner(t_arg *tmp, unsigned int n);
 void			out_to_16(long long n, char c);
-int             pointer_length(long long n);
+int       		pointer_length(unsigned int n);
 void			ft_put_n_char(char ch, int n);
-int             width_pointer(t_arg *tmp, long long n);
+int             width_pointer(t_arg *tmp, unsigned int n);
 int				processor(t_arg *tmp, va_list *va);
 
 #endif
